fix(review): split polling-indexer cli and restore examples target
