+++ crates/iota-sdk/examples/polling-indexer/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright (c) 2026 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod cli;
 pub mod config;
 pub mod db;
 pub mod indexer;
