+++ crates/iota-sdk/examples/polling-indexer/src/main.rs
@@ -3,7 +3,7 @@
 
 use clap::Parser;
 use iota_sdk::graphql_client::Client;
-use polling_indexer::{config::AppConfig, db, indexer::Indexer};
+use polling_indexer::{cli::Cli, config::AppConfig, db, indexer::Indexer};
 use tracing::info;
 use tracing_subscriber::{EnvFilter, fmt};
 
@@ -11,7 +11,7 @@ use tracing_subscriber::{EnvFilter, fmt};
 async fn main() -> anyhow::Result<()> {
     init_tracing();
 
-    let cli = polling_indexer::config::Cli::parse();
+    let cli = Cli::parse();
     let config = AppConfig::try_from(cli)?;
 
     let client = Client::new(&config.graphql_url)?;
